OOP Is A Mess
Stratigically attempting to correvtly initialize
and place my classes in the correct functions.

I might try to attempt initializing all classes from main only in a conditional way
depending on what other classes return to keep thongs better
organized and less confusing..

But if we are to follow the strict guidlines of oop
then we are are in a pickle.

But if I am to not do that, well then i will be sharing state between other classes with each other
and we are then again in the same pickle.

Oh well. Pickles are sometimes good